backend/requirements.txt
django==4.2.11
djangorestframework==3.15.0
Pillow==10.2.0
drf-nested-routers==0.93.5
djoser
djangorestframework-simplejwt==4.7.2